Builds with Xcode 8 and adapts a bit to new devices

+++ Classes/Settings/OutlineState.h
@@ -25,13 +25,13 @@
 
 @interface OutlineState : NSObject {
     OutlineSelectionType selectionType;
-    int selectedChildIndex;
+    long selectedChildIndex;
     int scrollPositionY;
     NSString *selectedLink;
 }
 
 @property (nonatomic) OutlineSelectionType selectionType;
-@property (nonatomic) int selectedChildIndex;
+@property (nonatomic) long selectedChildIndex;
 @property (nonatomic) int scrollPositionY;
 @property (nonatomic, retain) NSString *selectedLink;
